transform_mjit_header: Fix C99 test.
The test for pre-C99 compilers which don't recognize 'restrict' as a
keyword requires that the error message uses standard ASCII single
quotes.  But as of GCC 4.0, locale-specific quotes may be used, which
breaks the check, and in particular was causing some MacPorts builds
to fail.  The recommended approach is to run GCC with LC_ALL=C in any
case which expects to parse its messages.  This should be at worst
harmless with other compilers.

TESTED:
Previously failing MacPorts build using GCC 4.2 on OSX 10.5 now works,
as well as already successful non-MacPorts build.