Fix build on 10.4 x86.
The context register naming tweak was based on __DARWIN_UNIX03, which
isn't set correctly (at least by default).  Using the OS version macro
makes it work automatically.  Since it needs to be based on the SDK
version, it uses the MAX_ALLOWED form, though that requires including
AvailabilityMacros.h.  This is preferable to Availability.h, which
didn't exist in 10.4.

The behavior on iOS isn't entirely clear, but since this is only for
x86, that doesn't matter.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.6-10.15 x86_64,
and 11.x-12.x arm64.