Fix for fcopyfile() without COPYFILE_STATE_COPIED.
Versions of OSX prior to 10.6 implement fcopyfile() but not
the COPYFILE_STATE_COPIED function of copyfile_state_get().
Since ruby requires the latter in its use of the former, the
fix is to pretend that fcopyfile() doesn't exist when
COPYFILE_STATE_COPIED is undefined.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.6-10.15 x86_64,
and 11.x-12.x arm64.