Add *86-darwin* coroutine case.
The coroutine support uses ucontext as the generic mechanism when a
CPU-specific mechanism isn't provided.  But ucontext doesn't exist on
OSX prior to 10.5, leading to build errors on 10.4.  A CPU-specific
context handler ('x86') already exists, but wasn't being selected in
the darwin case.  This adds that case.

In addition to making builds possible under 10.4, this also switches
to the CPU-specific handler in later OS versions, which is more
consistent with the handling of other architectures.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.6-10.15 x86_64, and
11.x-12.x arm64.