Add ppc*-darwin8* coroutine case.
The coroutine code uses ucontext as the generic handler where
CPU-specific code is unavailable, but ucontext is unsupported
on OSX before 10.5, and no ppc-specific code is available,
so OSX builds failed on 10.4.  This adds a case to use the 'copy'
handler on OSX 10.4 (darwin8), while letting the later OSX versions
use ucontext as before.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.6-10.15 x86_64, and
11.x-12.x arm64.