.gdbinit: Avoid complete failure on older gdbs.
The "set startup-with-shell off" command was added to .gdbinit
for some later macOS versions, but it's not recognized by earlier
gdb versions, and the error aborts the .gdbinit processing.  This
moves the command to the end, so that other .gdbinit setup can
still happen in the error case.

It does not avoid the error message, which would require a more
elaborate fix.

TESTED:
The ruby-specific gdb additions are now set up correctly when the
new command is unrecognized.