Fix (build) crash on OSX <10.6.
Versions of OSX prior to 10.6 implement realpath(), but without the
extension that automatically allocates the target buffer when the
supplied pointer is NULL.  Since ruby expects that feature, the result
is a crash, and since the relevant code is used by miniruby in the
build procedure, this resulted in a crash during the build.

This change adds a wrapper to malloc() the buffer when needed, in the
relevant cases.  Otherwise, the wrapper is a macro that maps directly
to realpath().

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.6-10.15 x86_64, and
11.x-12.x arm64.