Add universal-darwin8* coroutine case.
Because the configure setup isn't CPU-specific in the universal
case, the coroutine selection needs to be the lowest common
denominator, which is 'copy' prior to OSX 10.5.  Later OSX
versions use ucontext as before.

This should fix universal builds on OSX 10.4, but dependency
issues prevent fully testing it.  It does get past the previous
10.4 universal build failure.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-10.15 x86_64, and
11.x-13.x arm64.  Universal builds tested on all of the above
except 10.4 (see above) and 10.14-10.15 (Apple restriction).