Patch configure to match source updates.
Patching configure avoids rerunning autoconf, which not only adds to
the build time but also has reproducibility issues.

Note that this patch must be applied after patching the sources, so
that make doesn't decide to rerun autoconf to fix the "outdated"
configure.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-10.15 x86_64, and
11.x-13.x arm64.