Fix bigdecimal build for macOS 12+.
The build of ext/bigdecimal/util relies on the use of the linker
option -undefined dynamic_lookup.  When targeting macOS 12 and later
(with Xcode 14 or later), the linker defaults to using chained fixups,
and warns that the -undefined dynamic_lookup option may no work
correctly in that case.  Although this is only a warning, and may not
even be applicable here, it has the effect of disabling the needed
option in the configure checks, resulting in a build failure.

Newer versions of Ruby have reworked the code such that that option
is no longer needed, but the change is to complicated to backport.
Adding the -no_fixup_chains option disables the new behavior and allows
the old method to work.

This means that the benefit of the new chained-fixup feature won't be
available with this version of Ruby, but that makes it no worse than
it was before.  It would be a bit better if this could be focused on the
one bundle that needs it, but that would be a more complicated change.

See the comment for additional info.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-10.15 x86_64, and
11.x-13.x arm64, including universal builds where applicable.