Fix build on OSX <10.6 with libunwind.h
MacPorts may provide libunwind.h via the libunwind-headers port, but
it only supports OSX 10.6 and later.  The presence of the header in
the unsupported case causes build errors.  This change adds an OS
version condition to the existing "Apple" case, avoiding the problem.

TESTED:
Built on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.6-12.x x86_64, and
11.x-14.x arm64.