Restore 3.3 version of linker warning handling.
The 3.3 configure setup detects whether certain linker warnings are
emitted, and if so arranges a wrapper to filter out the warnings.  The
3.4 version attempts to handle this by disabling the warnings, but
that mechanism doesn't work at all, resulting in invalid linker
options and eventually a build screwup.  So we revert to the old
wrapper approach.

This bug only appeared on one particular system where there was an
inconsistency between two frameworks that it was using for the test.

TESTED:
Now builds on all OS versions, including 10.13 with the mismatch.