Revert "CPPFLAGS is needed for dtrace not only INCFLAGS"
This reverts commit 060ae17c806f0ba955b1c5d825b890b6de6c47e3.

DTrace chokes on "-isystem" in CPPFLAGS, and the reverted commit
doesn't explain the need for CPPFLAGS.

If passing CPPFLAGS to DTrace is actually needed in some circumstance,
then it would need to use a modified version of CPPFLAGS that replaces
"-isystem" with "-I" in order to work correctly.  In the absence of
either evidence or an explanation for this need, we simply revert the
change.

TESTED:
Now builds successfully on macOS with dtrace enabled.